moved tests into folder
