+++ src/ProgressChart/__tests__/ProgressChart.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { ProgressChart } from './';
+import { ProgressChart } from '../';
 
 const TEST_DATA = [
   {
